+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 
 - Fix handling of `completionItem/resolve` when not all fields are present on the `CompletionItem`.
+- Fix handling of eager resolution of completions.
 
 ## 0.25.4